+++ frontend/app/[locale]/edit/metaTab.tsx
@@ -3,30 +3,19 @@ import Button, { buttonStyle } from "@/common/button.js";
 import Input from "@/common/input.js";
 import { checkYouTubeId, getYouTubeId } from "@/common/ytId.js";
 import { ChangeEvent, useEffect, useState } from "react";
-import msgpack from "@ygoe/msgpack";
-import { saveAs } from "file-saver";
-import {
-  ChartEdit,
-  ChartMin,
-  convertToMin,
-  lastIncompatibleVer,
-  validateChartMin,
-} from "@falling-nikochan/chart";
-import { getPasswd, setPasswd, unsetPasswd } from "@/common/passwdCache.js";
-import { addRecent } from "@/common/recent.js";
+import { ChartEdit, lastIncompatibleVer } from "@falling-nikochan/chart";
 import { initSession, SessionData } from "@/play/session.js";
 import { ExternalLink } from "@/common/extLink.js";
 import ProgressBar from "@/common/progressBar.js";
-import YAML from "yaml";
 import CheckBox from "@/common/checkBox.js";
 import Caution from "@icon-park/react/lib/icons/Caution.js";
 import { useTranslations } from "next-intl";
 import { HelpIcon } from "@/common/caption";
-import { luaExec } from "@falling-nikochan/chart/dist/luaExec";
 import { chartMaxEvent } from "@falling-nikochan/chart";
 import { useShareLink } from "@/common/shareLinkAndImage";
-import { isStandalone, updatePlayCountForReview } from "@/common/pwaInstall";
+import { isStandalone } from "@/common/pwaInstall";
 import { useRouter } from "next/navigation";
+import { useChartFile } from "./file";
 
 interface Props {
   chart?: ChartEdit;
@@ -148,7 +137,6 @@ interface Props2 {
 }
 export function MetaTab(props: Props2) {
   const t = useTranslations("edit.meta");
-  const te = useTranslations("error");
   const router = useRouter();
   const [errorMsg, setErrorMsg] = useState<string>("");
   const [saveMsg, setSaveMsg] = useState<string>("");
@@ -164,132 +152,33 @@ export function MetaTab(props: Props2) {
     props.chart.levels.length > 0 &&
     props.chart.levels.some((l) => l.notes.length > 0 && !l.unlisted);
 
+  const { remoteSave, remoteDelete, downloadExtension, localSave, load } =
+    useChartFile({
+      cid: props.cid,
+      chart: props.chart,
+      savePasswd: props.savePasswd,
+      currentPasswd: props.currentPasswd,
+      locale: props.locale,
+    });
   const save = async () => {
     setSaving(true);
-    const onSave = async (cid: string, changePasswd: string | null) => {
-      setErrorMsg(t("saveDone"));
-      props.setHasChange(false);
-      props.setConvertedFrom(props.chart!.ver);
-      if (changePasswd) {
-        if (props.savePasswd) {
-          try {
-            fetch(
-              process.env.BACKEND_PREFIX +
-                `/api/hashPasswd/${cid}?p=${changePasswd}`,
-              {
-                credentials:
-                  process.env.NODE_ENV === "development"
-                    ? "include"
-                    : "same-origin",
-              }
-            ).then(async (res) => {
-              setPasswd(cid, await res.text());
-            });
-          } catch {
-            //ignore
-          }
-        } else {
-          unsetPasswd(cid);
-        }
-        props.currentPasswd.current = changePasswd;
-      }
-    };
     props.chart!.changePasswd =
       props.newPasswd.length > 0 ? props.newPasswd : null;
-    if (props.cid === undefined) {
-      try {
-        const res = await fetch(
-          process.env.BACKEND_PREFIX + `/api/newChartFile`,
-          {
-            method: "POST",
-            body: msgpack.serialize(props.chart),
-            cache: "no-store",
-            credentials:
-              process.env.NODE_ENV === "development"
-                ? "include"
-                : "same-origin",
-          }
-        );
-        if (res.ok) {
-          try {
-            const resBody = (await res.json()) as {
-              message?: string;
-              cid?: string;
-            };
-            if (typeof resBody.cid === "string") {
-              props.setCid(resBody.cid);
-              history.replaceState(
-                null,
-                "",
-                `/${props.locale}/edit?cid=${resBody.cid}`
-              );
-              addRecent("edit", resBody.cid);
-              updatePlayCountForReview();
-              onSave(resBody.cid, props.chart!.changePasswd);
-            } else {
-              setErrorMsg(te("badResponse"));
-            }
-          } catch {
-            setErrorMsg(te("badResponse"));
-          }
-        } else {
-          try {
-            const message = ((await res.json()) as { message?: string })
-              .message;
-            if (te.has("api." + message)) {
-              setErrorMsg(te("api." + message));
-            } else {
-              setErrorMsg(message || te("unknownApiError"));
-            }
-          } catch {
-            setErrorMsg(te("unknownApiError"));
-          }
-        }
-      } catch (e) {
-        console.error(e);
-        setErrorMsg(te("api.fetchError"));
-      }
+    const result = await remoteSave();
+    if (result.isError) {
+      setErrorMsg(result.message || "");
     } else {
-      const q = new URLSearchParams();
-      if (props.currentPasswd.current) {
-        q.set("p", props.currentPasswd.current);
-      } else if (getPasswd(props.cid)) {
-        q.set("ph", getPasswd(props.cid)!);
-      }
-      try {
-        const res = await fetch(
-          process.env.BACKEND_PREFIX +
-            `/api/chartFile/${props.cid}?` +
-            q.toString(),
-          {
-            method: "POST",
-            body: msgpack.serialize(props.chart),
-            cache: "no-store",
-            credentials:
-              process.env.NODE_ENV === "development"
-                ? "include"
-                : "same-origin",
-          }
-        );
-        if (res.ok) {
-          props.setHasChange(false);
-          onSave(props.cid, props.chart!.changePasswd);
-        } else {
-          try {
-            const message = ((await res.json()) as { message?: string })
-              .message;
-            if (te.has("api." + message)) {
-              setErrorMsg(te("api." + message));
-            } else {
-              setErrorMsg(message || te("unknownApiError"));
-            }
-          } catch {
-            setErrorMsg(te("unknownApiError"));
-          }
-        }
-      } catch (e) {
-        console.error(e);
-        setErrorMsg(te("api.fetchError"));
+      setErrorMsg(result.message || "");
+      props.setCid(result.cid!);
+      history.replaceState(
+        null,
+        "",
+        `/${props.locale}/edit?cid=${result.cid!}`
+      );
+      props.setHasChange(false);
+      props.setConvertedFrom(props.chart!.ver);
+      if (props.chart!.changePasswd) {
+        props.currentPasswd.current = props.chart!.changePasswd;
       }
     }
     props.chart!.changePasswd = null;
@@ -305,113 +194,28 @@ export function MetaTab(props: Props2) {
         break;
       }
     }
-    const q = new URLSearchParams();
-    if (props.currentPasswd.current) {
-      q.set("p", props.currentPasswd.current);
-    } else if (getPasswd(props.cid!)) {
-      q.set("ph", getPasswd(props.cid!)!);
-    }
-    try {
-      const res = await fetch(
-        process.env.BACKEND_PREFIX +
-          `/api/chartFile/${props.cid}?` +
-          q.toString(),
-        {
-          method: "DELETE",
-          cache: "no-store",
-          credentials:
-            process.env.NODE_ENV === "development" ? "include" : "same-origin",
-        }
-      );
-      if (res.ok) {
-        if (isStandalone()) {
-          history.back();
-        } else {
-          window.close();
-        }
-      }
-    } catch (e) {
-      console.error(e);
-    }
+    await remoteDelete();
   };
-  const downloadExtension = `fn${props.chart?.ver}.yml`;
   const download = () => {
-    const yml = YAML.stringify(convertToMin(props.chart!), {
-      indentSeq: false,
-    });
-    const filename = `${props.cid}_${props.chart?.title}.${downloadExtension}`;
-    saveAs(new Blob([yml]), filename);
-    setSaveMsg(`${t("saveDone")} (${filename})`);
+    const result = localSave();
+    setSaveMsg(result.message || "");
   };
   const upload = async (e: ChangeEvent) => {
     const target = e.target as HTMLInputElement;
     if (target.files && target.files.length >= 1) {
       const f = target.files[0];
       const buffer = await f.arrayBuffer();
-      setUploadMsg("");
-      let originalVer: number = 0;
-      let newChart: ChartEdit | null = null;
-      try {
-        const content: ChartMin = YAML.parse(new TextDecoder().decode(buffer));
-        if (typeof content.ver === "number") {
-          originalVer = content.ver;
-        }
-        const newChartMin = await validateChartMin(content);
-        newChart = {
-          ...newChartMin,
-          changePasswd: null,
-          published: false,
-          levels: await Promise.all(
-            newChartMin.levels.map(async (l) => ({
-              ...l,
-              ...(
-                await luaExec(
-                  process.env.ASSET_PREFIX + "/assets/wasmoon_glue.wasm",
-                  l.lua.join("\n"),
-                  false
-                )
-              ).levelFreezed,
-            }))
-          ),
-        };
-      } catch (e1) {
-        console.warn("fallback to msgpack deserialize");
-        try {
-          const content: ChartMin = msgpack.deserialize(buffer);
-          if (typeof content.ver === "number") {
-            originalVer = content.ver;
-          }
-          const newChartMin = await validateChartMin(content);
-          newChart = {
-            ...newChartMin,
-            changePasswd: null,
-            published: false,
-            levels: await Promise.all(
-              newChartMin.levels.map(async (l) => ({
-                ...l,
-                ...(
-                  await luaExec(
-                    process.env.ASSET_PREFIX + "/assets/wasmoon_glue.wasm",
-                    l.lua.join("\n"),
-                    false
-                  )
-                ).levelFreezed,
-              }))
-            ),
-          };
-        } catch (e2) {
-          console.error(e1);
-          console.error(e2);
-          setUploadMsg(t("loadFail"));
-        }
-      }
-      if (newChart) {
-        if (confirm(t("confirmLoad"))) {
-          props.setChart(newChart);
-          props.setConvertedFrom(originalVer);
-        }
-      }
       target.value = "";
+      setUploadMsg("");
+      const result = await load(buffer);
+      if (result.isError) {
+        setUploadMsg(result.message || "");
+      } else {
+        if (confirm(t("confirmLoad"))) {
+          props.setChart(result.chart!);
+          props.setConvertedFrom(result.originalVer!);
+        }
+      }
     }
   };
